backend/requirements.txt
fastapi~=0.128.0
sqlmodel
pydantic-settings
asyncpg==0.29.0
greenlet
uvicorn[standard]
python-multipart
aiofiles
